Refactor profiler
