+++ packages/profiler/src/main.rs
@@ -1,3 +1,4 @@
+mod code_blocks;
 mod middleware;
 mod operators;
